[Refactor] #651 - SoptampFlow의 코디네이터를 하나로 통합
